+++ src/main/java/DTO/Bokning.java
@@ -1,3 +1,5 @@
+package DTO;
+
 
 import java.util.List;
