+++ src/main/java/DTO/Pass.java
@@ -1,3 +1,5 @@
+package DTO;
+
 import java.util.Date;
 
 public class Pass {
