+++ src/main/java/DTO/Test.java
@@ -1,3 +1,5 @@
+package DTO;
+
 public class Test {    
     public static void main(String[] args) {
