+++ src/main/java/DTO/Tidslucka.java
@@ -1,3 +1,5 @@
+package DTO;
+
 
 import java.sql.Time;
